Create a new `frontAndBack` from `from` and `back` patches when a `TRI` or `POLY` mesh is generated
This is done to simplify boundary fields in `0/`. It is simpler and less
redundant to have only `frontAndBack` instead of `front`, `back, and `frontAndBack` at the same time, or even `"front|back|frontAndBack`.

For that `createPatch` is used to generate the `frontAndBack` patch as
the last step in generating a `TRI` or `POLY` mesh.