oscillatingSphereInCube/base/0.orig/pointDisplacement: Use the `oscillatingLinearMotion` function as an alternative implementation to the previous `codedFixedValue`
In this approach, the reference displacement function, originally
expressed as a sine function, is rewritten as a cosine function with a
phase shift.

x(t) = -h*\cos(2*\pi*t) + h
     = -h*\sin(2*\pi*t + \pi/2) + h
     = -h*\sin(2*\pi*(t + (\pi/2)/(2\pi))) + h
     = -h*\sin(2*\pi*(t + 1/4)) + h

Finally, pattern matching, against the formula for the
`oscillatingLinearMotion` function, is applied to extract the necessary
parameters.

x =  A*\sin(B*(t + C)) + D